Move search resolver logic to separate file
This allows us to better encapsulate the resolution behavior from the
GraphQL interface/schema definition. We'll be able to add on more
optimizations with this split in place.
